feat(decl/process): add process builder
